Removed compiled code. Restructured UI class to be in a ui package
